requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
Jinja2
python-multipart
pydantic_settings
psycopg2